[Fix] Multiple selection on radio button
This fix doesn't guarantee to fix the issue because unable to be simulated consistently and unable to be simulated by me as well. This fix changed internal implementation into safer approach, safer in a sense that did not hold CheckBox/RadioButton in the field. Also synchronizing/locking selected ids field to avoid inconsistent state when accessing the field